services/user_management/requirements.txt
# requirements.txt
fastapi
uvicorn
pydantic
python-jose
passlib[bcrypt]
psycopg2-binary
